playlist store module renaming
