+++ functions/client/src/store/index.js
@@ -1,6 +1,6 @@
 import Vuex from 'vuex';
 import Vue from 'vue';
-import Playlist from './modules/PlaylistAudio';
+import Playlist from './modules/Playlist';
 import Search from './modules/Search';
 import Auth from './modules/Auth';
 Vue.use(Vuex);
